move unit tests to matching directory
